+++ .ci/install_linux.sh
@@ -1,2 +1,2 @@
-cp "${TRAVIS_BUILD_DIR}" src
+cp -r "${TRAVIS_BUILD_DIR}" src
 ./scripts/internal-distro.py --workspace=src distribution.yml --repository "${REPOSITORY}"